introduce `MaxCommission`
